PyrMapperPrim.cpp: added Mapper::Signal::input_handler()
This is the callback that gets called on an incoming signal update.
also added MapperSignal:prCallAction().

input_handler() then calls MapperSignal:prCallAction() on the
corresponding MapperSignal instance.  Then prCallAction() calls the
function stored in the local action in the MapperSignal instance.

A pointer to the MapperSignal instance that gets invoked is stored in
the signal's user_data.