PyrMapperPrim: add error handling for null devstructs
check for a null devstruct pointer in the mapper primitive functions.
this would typically indicate that the device has already been freed.

For the functions that are internal to the lang-side implementation the
end-users won't be using, we're returning errFailed, which will dump a
stack trace in sclang.  This will enforce error checking in the
MapperDevice implementation.

The end-user facing functions are a bit more forgiving, issuing a
warning and maybe returning nil.